refactor: move TestWrapper to utils/tests
