+++ packages/outstatic/src/components/tests/delete-document-button.test.tsx
@@ -1,5 +1,5 @@
 import { DeleteDocumentButton } from '@/components/delete-document-button'
-import { TestWrapper } from '@/utils/TestWrapper'
+import { TestWrapper } from '@/utils/tests/test-wrapper'
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 
 // Mock the useOstSession hook
